minimum bug fixed to start Hal

+++ storm_control/hal4000/focusLock/lockDisplay.py
@@ -199,14 +199,14 @@ class QCamDisplay(QtWidgets.QWidget):
         painter.drawRect(0, 0, w, w)
 
         # Draw image.
-        y_start = self.display_pixmap.height()/2 - self.camera_image.height()/2
+        y_start = int(self.display_pixmap.height()/2 - self.camera_image.height()/2)
         destination_rect = QtCore.QRect(0, y_start, self.camera_image.width(), self.camera_image.height())
         painter.drawImage(destination_rect, self.camera_image)
 
         # Draw bounding rectangle.
         if (w != h):
             pen = QtGui.QPen(QtGui.QColor(255,0,0))
-            pen.setWidth(self.display_pixmap.width()/self.width())
+            pen.setWidth(int(self.display_pixmap.width()/self.width()))
             painter.setPen(pen)
             painter.setBrush(QtGui.QColor(0,0,0,0))
             painter.drawRect(destination_rect)
@@ -215,7 +215,9 @@ class QCamDisplay(QtWidgets.QWidget):
         if (self.zoom_im_x >= 0):
             self.zoom_image = QtGui.QImage(self.zoom_size, self.zoom_size, QtGui.QImage.Format_RGB32)
             painter = QtGui.QPainter(self.zoom_image)
-            painter.drawPixmap(0, 0, self.display_pixmap, self.zoom_im_x, self.zoom_im_y, self.zoom_size, self.zoom_size)
+            painter.drawPixmap(0, 0, self.display_pixmap, 
+                               int(self.zoom_im_x), int(self.zoom_im_y), 
+                               int(self.zoom_size), int(self.zoom_size))
         else:
             self.zoom_image = False
 
@@ -315,12 +317,13 @@ class QCamDisplay(QtWidgets.QWidget):
             if self.adjust_mode:
 
                 painter.setPen(QtGui.QColor(100,100,100))
-                painter.drawLine(0.0, 0.5*self.height(), self.width(), 0.5*self.height())
+                painter.drawLine(0, int(0.5*self.height()), int(self.width()), int(0.5*self.height()))
                 for mult in [0.25, 0.5, 0.75]:
-                    painter.drawLine(mult*self.width(), 0.0, mult*self.width(), self.height())
+                    painter.drawLine(int(mult*self.width()), 0, int(mult*self.width()), int(self.height()))
 
                 if self.zoom_image:
-                    destination_rect = QtCore.QRect(self.zoom_x, self.zoom_y, self.zoom_size, self.zoom_size)
+                    destination_rect = QtCore.QRect(int(self.zoom_x), int(self.zoom_y), 
+                                                    int(self.zoom_size), int(self.zoom_size))
                     painter.drawImage(destination_rect, self.zoom_image)
                     painter.setPen(QtGui.QColor(200,200,200))
                     painter.setBrush(QtGui.QColor(0,0,0,0))
